pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "my-chat-ai-comfyui"
version = "0.1.0"
description = "A powerful integration system that bridges conversational AI interfaces with ComfyUI's visual programming environment"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "tqer39", email = "tqer39@gmail.com"},
]
keywords = ["ai", "comfyui", "chat", "image-generation", "stable-diffusion"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "requests>=2.31.0",
    "websocket-client>=1.6.0",
    "aiohttp>=3.8.0",
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "onnxruntime>=1.15.0",
    "openai>=1.0.0",
    "discord.py>=2.3.0",
    "slack-sdk>=3.21.0",
    "python-telegram-bot>=20.0",
    "spacy>=3.6.0",
    "nltk>=3.8.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "sqlalchemy>=2.0.0",
    "loguru>=0.7.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "tqdm>=4.65.0",
    "cryptography>=41.0.0",
    "httpx>=0.24.0",
    "python-dateutil>=2.8.0",
    "asyncio>=3.4.3",
    "aiofiles>=23.0.0",
    "imageio>=2.31.0",
    "huggingface-hub>=0.16.0",
    "jsonschema>=4.19.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.6",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
    "sphinx>=7.1.0",
    "sphinx-rtd-theme>=1.3.0",
]

[project.urls]
Homepage = "https://github.com/tqer39/my-chat-ai-comfyui"
Repository = "https://github.com/tqer39/my-chat-ai-comfyui"
Documentation = "https://github.com/tqer39/my-chat-ai-comfyui/docs"
"Bug Tracker" = "https://github.com/tqer39/my-chat-ai-comfyui/issues"

[project.scripts]
chat-ai-comfyui = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E203"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
explicit_package_bases = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "setup.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
